requirements.txt
# ============================================
# Smart Document Insights - Python Dependencies
# ============================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# --- Database ---
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0

# --- AI / LLM ---
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7

# --- Embeddings & Vector Store ---
sentence-transformers==3.3.0
faiss-cpu==1.9.0
numpy==1.26.4

# --- Document Processing ---
PyMuPDF==1.24.14
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0

# --- Text Processing ---
chardet==5.2.0
langdetect==1.0.9

# --- Background Tasks ---
celery[redis]==5.4.0
redis==5.2.1

# --- Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# --- Utilities ---
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0

# --- Logging ---
structlog==24.4.0

# --- Testing ---
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
